requirements.txt
Flask==3.1.0
tensorflow==2.19.0
gunicorn==20.1.0
Pillow==11.1.0
gdown==5.2.0
numpy==1.24.1

